requirements-scheduler.txt
Jinja2==3.1.1
google-cloud-bigquery==3.0.1
pytest==6.2.5
